[IR] Change kotlin-like IR dump format of value parameters and arguments
and migrate the code to new parameter API.

Extension receiver and context parameters are now rendered alongside
regular arguments, because they are just synthetic sugar for them,
but are commented-out, because it is not how they are really specified
in kotlin syntax.

Previously extension receiver was actually rendered "normally"
(the same way as dispatch receiver). That way is closer to source code,
but has two problems:
- Dispatch and extension receivers are no distinguishable.
- A situation when both dispatch and extension receivers are present is
not representable that way, and requires some hack to render both.

#KT-74275 Fixed
#KT-73220 Fixed